IMPROVE: code separation
